docs: fix flake part

+++ bin/nvchad.sh
@@ -46,7 +46,9 @@ wrapper() {
 main() {
     check_init
     if ! [ -s "${_CONF_DIR}/lazy-lock.json" ]; then
-      rm "$_CONF_DIR/lazy-lock.json" 
+      if [ -e "${_CONF_DIR}/lazy-lock.json" ]; then
+        rm "$_CONF_DIR/lazy-lock.json" 
+      fi
     fi
     wrapper "$@"
 }
